+++ Mburu-Elvis/README.md
@@ -1,0 +1,41 @@
+# Demo Project
+
+This is a project to implement a simple calculator
+
+## The operations supported by the Calculator
+
+### 1. Addition
+
+This operation carries out summation of two operands
+
+This operation is represented by the operator `+`
+
+### 2. Subtraction
+
+The calculator finds the difference between two numbers
+
+This operation is represented by the operator `-`
+
+### 3. Multiplication
+
+The calculator finds the product between two numbers
+
+This operation is represented by the operator `*`
+
+### 4. Exponentiation
+
+The calculator perform exponential multiplication
+
+This operation is represented by the operator `**`
+
+### 5. Division
+
+The calculator finds the quotient after division of two numbers
+
+This operation is represented by the operator `/`
+
+### 6. Modulo Division
+
+The calculator finds the remainder between the division of a dividend and quotient
+
+This operation is represented by the operator `%`
