+++ Mburu-Elvis/calculations.c
@@ -1,7 +1,12 @@
 #include "calc.h"
-#include <math.h>
-#include <stdlib.h>
 
+/**
+ * add1 - a function to perform addition
+ * @a: The first operand
+ * @b: The second operand
+ *
+ * Return: The sum of the two operands
+*/
 float add1(float a, float b)
 {
     float result = 0;
@@ -10,6 +15,13 @@ float add1(float a, float b)
     return (result);
 }
 
+/**
+ * sub1 - a function to perform subtraction
+ * @a: The first operand
+ * @b: The second operand
+ *
+ * Return: The difference between the operands
+*/
 float sub1(float a, float b)
 {
     float result = 0;
@@ -18,6 +30,13 @@ float sub1(float a, float b)
     return (result);
 }
 
+/**
+ * mul1 - a function to perform multiplication
+ * @a: The first operand
+ * @b: The second operand
+ *
+ * Return: The product of the two operands
+*/
 float mul1(float a, float b)
 {
     float result = 1;
@@ -26,24 +45,14 @@ float mul1(float a, float b)
     return (result);
 }
 
+/**
+ * div1 - a function to perform division
+ * @a: The Dividend
+ * @b: The Divisor
+ * 
+ * Return: The quotient from the division
+*/
 float div1(float a, float b)
 {
     return ((a) * (b));
-}
-
-float mod1(float a, float b)
-{
-    float result = 1;
-
-    result = fmod(a, b);
-    return (result);
-}
-
-float exp1(float a, float b)
-{
-    float result = 1, i;
-
-    for (i = 0; i < b; i++)
-        result *= a;
-    return (result);
 }
